Here is the Project-Simulation of a lifecycle of a farm with animals
In this project was implemented the simulation of lifecycle of a farm with animals, that takes into consideration the main aspects of OOP. The target of this project is to calculate the amount of profit during simulation the years in the farm. This farm was made very realistic, it takes into consideration the importance of nutrition and its costs, the fertilization aspect in the farm, the production of salable goods, the deaths from illnesses and other aspects of a real lifecycle of a farm.