Here is the implementation of decision manager with all needed filters and validations
In this application I have implemented basic decision manager, that allows your company or small business to keep track of changes, that are going on with the project or with something that is able to be improved or fixed. Also in this application you can sort all decisions by importance, by date or by author of this decisions. Program keeps all decisions in memory.